+++ common/app-starters-security-common/src/main/java/org/springframework/cloud/stream/app/security/common/AppStarterWebFluxSecurityAutoConfiguration.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.stream.app.security.common;
+
+import org.springframework.boot.actuate.autoconfigure.security.reactive.ReactiveManagementWebSecurityAutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.boot.autoconfigure.security.reactive.ReactiveSecurityAutoConfiguration;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
+import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.web.server.SecurityWebFilterChain;
+import org.springframework.security.web.server.WebFilterChainProxy;
+import org.springframework.web.reactive.config.WebFluxConfigurer;
+
+import reactor.core.publisher.Flux;
+
+/**
+ * @author Artem Bilan
+ *
+ * @since 3.0
+ */
+@Conditional(OnHttpCsrfOrSecurityDisabled.class)
+@Configuration
+@ConditionalOnClass({ Flux.class, EnableWebFluxSecurity.class, WebFilterChainProxy.class, WebFluxConfigurer.class })
+@ConditionalOnMissingBean(WebFilterChainProxy.class)
+@ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.REACTIVE)
+@AutoConfigureBefore(value = { ReactiveManagementWebSecurityAutoConfiguration.class,
+		ReactiveSecurityAutoConfiguration.class })
+@EnableConfigurationProperties(AppStarterWebSecurityAutoConfigurationProperties.class)
+public class AppStarterWebFluxSecurityAutoConfiguration {
+
+	@Bean
+	public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http,
+			AppStarterWebSecurityAutoConfigurationProperties securityProperties) {
+		if (!securityProperties.isCsrfEnabled()) {
+			http.csrf().disable();
+		}
+		if (!securityProperties.isEnabled()) {
+			http.authorizeExchange()
+					.anyExchange()
+					.permitAll();
+		}
+		return http.build();
+	}
+
+}
