+++ common/app-starters-security-common/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.stream.app.security.common.AppStarterWebSecurityAutoConfiguration
+org.springframework.cloud.stream.app.security.common.AppStarterWebSecurityAutoConfiguration,\
+org.springframework.cloud.stream.app.security.common.AppStarterWebFluxSecurityAutoConfiguration
